% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{evolve}
\alias{evolve}
\title{Let the individuals of a population move around}
\usage{
evolve(population, n, prop)
}
\arguments{
\item{population}{Population to transform}

\item{n}{Number of individuals to introduce}

\item{prop}{Proportion of individuals to remove}
}
\value{
A population
}
\description{
Let the individuals of a population move around
}
