% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{generate_population}
\alias{generate_population}
\title{Generate a new population}
\usage{
generate_population(n_ind, size_hh)
}
\arguments{
\item{n_ind}{Number of individuals}

\item{size_hh}{Average size of households}
}
\value{
A population
}
\description{
Generate a new population
}
