% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_generator.R
\name{hh_geometric}
\alias{hh_geometric}
\title{Construct a basic household generator}
\usage{
hh_geometric(mean_size)
}
\arguments{
\item{mean_size}{Average size of households}
}
\value{
A household generator
}
\description{
Construct a basic household generator
}
