% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhsim.R
\docType{package}
\name{hhsim-package}
\alias{hhsim}
\alias{hhsim-package}
\title{hhsim: Household Population Simulation}
\description{
hhsim allows to simulate some household dynamics by allowing them to
merge, split and exit.
}
\author{
\strong{Maintainer}: Jean Rubin \email{john.doe@test.com}

}
