% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{introduce}
\alias{introduce}
\title{Introduce new individuals in a population}
\usage{
introduce(population, n, size_hh)
}
\arguments{
\item{population}{Population to transform}

\item{n}{Number of individuals to introduce}

\item{size_hh}{Average size of introduced households}
}
\value{
A population
}
\description{
Introduce new individuals in a population
}
