% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{remove}
\alias{remove}
\title{Remove individuals from a population}
\usage{
remove(population, n, prop)
}
\arguments{
\item{population}{Population to transform}

\item{n}{Number of individuals to remove}

\item{prop}{Proportion of individuals to remove}
}
\value{
A population
}
\description{
Remove individuals from a population
}
