% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{generate_population}
\alias{generate_population}
\title{Generate a new population}
\usage{
generate_population(n_ind, hh_gen)
}
\arguments{
\item{n_ind}{Number of individuals}

\item{hh_gen}{A household generator}
}
\value{
A population
}
\description{
Generate a new population
}
